🚚 Move FileDetail and SelectElement component to /documents/components
